Get working with ghc-8.2
inline-c has to be limited to 0.5.6.1 because inline-c-0.6.0.0 uses
TH.addForeignFile, which doesn't seem to support separate c++ flags,
which is a problem because Halide requires c++11. Passing -std=c++11 to
to the c-flags (via -optc) does let halide compile but the runtime fails
to compile on MacOS because it complains about using --std=c++11.

There is a proposal that would help solve this problem on Trac:
  https://ghc.haskell.org/trac/ghc/ticket/14298
which allows passing a bytestring containing the object file. However it
looks like this would make it into ghc-8.6 at the earliest.